Adding Background Music + Obstacle Collision Sound
BGM has been added to the game. Its called Winning by AshamaluevMusic. It is copyright free. For this, an empty object was created and named "Main Music", in which the component "Audio Source" was added, our audio is selected in the "Audio Clip" section.
For the Obstacle_Hit_Sound, the "Audio Source" component is added under the Player object itself, followed by slight modification in the PlayerCollision.cs script so that the sound is played when the Player collides with an Obstacle.